+++ src/utils/icons.js
@@ -1,0 +1,23 @@
+import partly_cloudy_day from './img/cloudy-day-3.svg';
+import partly_cloudy_night from './img/cloudy-night-3.svg';
+import cloudy from './img/cloudy.svg';
+import rain from './img/rainy-6.svg';
+import clear_day from './img/day.svg';
+import clear_night from './img/night.svg';
+import snow from './img/snowflake.svg';
+import sleet from './img/snowy-6.svg';
+import weather from './img/weather.svg';
+
+export const icons = {
+  'partly-cloudy-day': partly_cloudy_day,
+  'partly-cloudy-night': partly_cloudy_night,
+  'clear-day': clear_day,
+  'clear-night': clear_night,
+  'rain': rain,
+  'snow': snow,
+  'sleet': sleet,
+  'wind': '',
+  'fog': '',
+  'cloudy': cloudy,
+  'default': weather,
+}
